+++ src/database/operations/changemakerFieldValues/createChangemakerFieldValue.ts
@@ -1,0 +1,26 @@
+import { generateCreateOrUpdateItemOperation } from '../generators';
+import { decorateWithFileDownloadUrl } from '../../../decorators/changemakerFieldValue';
+import type {
+	ChangemakerFieldValue,
+	InternallyWritableChangemakerFieldValue,
+} from '../../../types';
+
+const createChangemakerFieldValue = generateCreateOrUpdateItemOperation<
+	ChangemakerFieldValue,
+	InternallyWritableChangemakerFieldValue,
+	[]
+>(
+	'changemakerFieldValues.insertOne',
+	[
+		'changemakerId',
+		'baseFieldShortCode',
+		'batchId',
+		'value',
+		'isValid',
+		'goodAsOf',
+	],
+	[],
+	decorateWithFileDownloadUrl,
+);
+
+export { createChangemakerFieldValue };