+++ src/routers/index.ts
@@ -9,6 +9,8 @@ import { opportunitiesRouter } from './opportunitiesRouter';
 import { organizationsRouter } from './organizationsRouter';
 import { changemakersRouter } from './changemakersRouter';
 import { changemakerProposalsRouter } from './changemakerProposalsRouter';
+import { changemakerFieldValueBatchesRouter } from './changemakerFieldValueBatchesRouter';
+import { changemakerFieldValuesRouter } from './changemakerFieldValuesRouter';
 import { platformProviderResponsesRouter } from './platformProviderResponsesRouter';
 import { proposalsRouter } from './proposalsRouter';
 import { proposalVersionsRouter } from './proposalVersionsRouter';
@@ -26,6 +28,11 @@ rootRouter.use('/baseFields', baseFieldsRouter);
 rootRouter.use('/tasks', tasksRouter);
 rootRouter.use('/changemakers', changemakersRouter);
 rootRouter.use('/changemakerProposals', changemakerProposalsRouter);
+rootRouter.use(
+	'/changemakerFieldValueBatches',
+	changemakerFieldValueBatchesRouter,
+);
+rootRouter.use('/changemakerFieldValues', changemakerFieldValuesRouter);
 rootRouter.use('/dataProviders', dataProvidersRouter);
 rootRouter.use('/files', filesRouter);
 rootRouter.use('/funders', fundersRouter);