+++ src/test/factories/index.ts
@@ -1,2 +1,3 @@
+export * from './createTestBaseField';
 export * from './createTestFile';
 export * from './createTestS3Bucket';