+++ src/types/index.ts
@@ -9,6 +9,8 @@ export * from './BulkUploadLog';
 export * from './BulkUploadTask';
 export * from './Bundle';
 export * from './Changemaker';
+export * from './ChangemakerFieldValue';
+export * from './ChangemakerFieldValueBatch';
 export * from './ChangemakerProposal';
 export * from './CheckResult';
 export * from './CopyBaseFieldsJobPayload';